third_party/blink/web_tests/fast/css/counters/counter-increment-tests-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.counterValue(document.getElementById('test3')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test4')) is '-2147483647'
PASS internals.counterValue(document.getElementById('test5')) is '0'
PASS internals.counterValue(document.getElementById('test6')) is '0'
PASS internals.counterValue(document.getElementById('test7')) is '0'
PASS internals.counterValue(document.getElementById('test8')) is '10'
PASS internals.counterValue(document.getElementById('test9')) is '10'
PASS internals.counterValue(document.getElementById('test10')) is '2147483646'
PASS internals.counterValue(document.getElementById('test11')) is '2147483647'
PASS internals.counterValue(document.getElementById('test12')) is '2147483647'
PASS internals.counterValue(document.getElementById('test13')) is '2147483647'
PASS internals.counterValue(document.getElementById('test14')) is '2'
PASS internals.counterValue(document.getElementById('test15')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test16')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test17')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test18')) is '0'
PASS internals.counterValue(document.getElementById('test19')) is '0'
PASS internals.counterValue(document.getElementById('test20')) is '0'
PASS internals.counterValue(document.getElementById('test21')) is '20'
PASS internals.counterValue(document.getElementById('test22')) is '20'
PASS internals.counterValue(document.getElementById('test23')) is '2147483647'
PASS internals.counterValue(document.getElementById('test24')) is '2147483647'
PASS internals.counterValue(document.getElementById('test25')) is '2147483647'
PASS internals.counterValue(document.getElementById('test26')) is '2147483647'
PASS internals.counterValue(document.getElementById('test27')) is '3'
PASS internals.counterValue(document.getElementById('test28')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test29')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test30')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test31')) is '0'
PASS internals.counterValue(document.getElementById('test32')) is '0'
PASS internals.counterValue(document.getElementById('test33')) is '0'
PASS internals.counterValue(document.getElementById('test34')) is '30'
PASS internals.counterValue(document.getElementById('test35')) is '30'
PASS internals.counterValue(document.getElementById('test36')) is '2147483647'
PASS internals.counterValue(document.getElementById('test37')) is '2147483647'
PASS internals.counterValue(document.getElementById('test38')) is '2147483647'
PASS internals.counterValue(document.getElementById('test39')) is '2147483647'
PASS internals.counterValue(document.getElementById('test40')) is '32'
PASS internals.counterValue(document.getElementById('test41')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test42')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test43')) is '-2147483648'
PASS internals.counterValue(document.getElementById('test44')) is '0'
PASS internals.counterValue(document.getElementById('test45')) is '0'
PASS internals.counterValue(document.getElementById('test46')) is '0'
PASS internals.counterValue(document.getElementById('test47')) is '320'
PASS internals.counterValue(document.getElementById('test48')) is '320'
PASS internals.counterValue(document.getElementById('test49')) is '2147483647'
PASS internals.counterValue(document.getElementById('test50')) is '2147483647'
PASS internals.counterValue(document.getElementById('test51')) is '2147483647'
PASS internals.counterValue(document.getElementById('test52')) is '2147483647'
PASS internals.counterValue(document.getElementById('test53')) is '5'
PASS internals.counterValue(document.getElementById('none53')) is '5'
PASS internals.counterValue(document.getElementById('test53a')) is '10'
PASS internals.counterValue(document.getElementById('wrapper54')) is '5'
PASS internals.counterValue(document.getElementById('test54')) is '10'
PASS internals.counterValue(document.getElementById('test55')) is '1'
PASS internals.counterValue(document.getElementById('test56')) is '-2147483648'
